Just refactoring a tiny bit to make it less ridiculous.
