+++ src/hypermedia/mod.rs
@@ -1,2 +1,3 @@
 pub mod router;
+mod schema;
 mod service;
