+++ src/hypermedia/schema/auth.rs
@@ -1,0 +1,4 @@
+pub struct MailToUser {
+    pub email: Option<String>,
+    pub verification_code: Option<String>,
+}
